Move errors to utils folders
